frontend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Session==0.5.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
Flask-Babel==4.0.0
Flask-Mail==0.9.1
Flask-SocketIO==5.3.5
python-socketio==5.10.0

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7
bcrypt==4.1.2
python-dotenv==1.0.0
werkzeug==3.0.1

# HTTP & API
requests==2.31.0
urllib3==2.1.0
httpx==0.25.2

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# Forms & Validation
WTForms==3.1.1
email-validator==2.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2
xlsxwriter==3.1.9

# Charts & Visualization
plotly==5.18.0
matplotlib==3.8.2

# PDF Generation
reportlab==4.0.7
weasyprint==60.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
faker==20.1.0

# Development
python-decouple==3.8
gunicorn==21.2.0
eventlet==0.33.3

# Monitoring
prometheus-flask-exporter==0.23.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pycountry==23.12.11
phonenumbers==8.13.27